requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
idna==3.3
joblib==1.1.0
markovify==0.9.3
nltk==3.7
praw==7.5.0
prawcore==2.3.0
python-dotenv==0.19.2
regex==2022.1.18
requests==2.27.1
schedule==1.1.0
textblob==0.17.1
tqdm==4.62.3
Unidecode==1.3.3
update-checker==0.18.0
urllib3==1.26.8
websocket-client==1.3.1
